Remove global scope access
I want to include this polyfill in polyfill.io.  It's good practice for polyfills to use `this` for the global reference and attach things to it rather than implicitly reading globals, so we'd very much appreciate this change in order to be happier having the polyfill in our service.  Another good effect of this change is that in CommonJS and AMD environments, you will no longer leak the global.